+++ tutor/scripts.py
@@ -9,8 +9,7 @@ def migrate(root, run_func):
     config = tutor_config.load(root)
 
     click.echo(fmt.info("Creating all databases..."))
-    # Note: run this only when lms is activated?
-    run_template(root, config, "lms", "create_databases.sh", run_func)
+    run_template(root, config, "mysql-client", "create_databases.sh", run_func)
 
     if config["ACTIVATE_LMS"]:
         click.echo(fmt.info("Running lms migrations..."))