+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Bugfix] Fix installing a locally cloned requirement repository
 - [Improvement] Add `--no-cache` option to `images build`
 - [Improvement] Make it possible to configure the notes service hostname
+- [Improvement] Better, more robust MySQL initialization
 
 ## 3.3.10 (2019-05-15)